pyproject.toml
[project]
name = "mcp-digimon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastmcp>=3.1.0",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.11",
    "starlette>=0.52.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
